refactor(hooks): extract event countdown hook into separate file
